[working] Refactoring complete

+++ msg/client/stream_test.go
@@ -12,9 +12,9 @@ func TestStream_init(t *testing.T) {
 	assert.Equal(t, []byte{}, buf.Data())
 	tail, cur, head, _ := buf.Snapshot()
 
-	assert.Equal(t, uint32(0), tail.offset)
-	assert.Equal(t, uint32(0), cur.offset)
-	assert.Equal(t, uint32(0), head.offset)
+	assert.Equal(t, uint64(0), tail.offset)
+	assert.Equal(t, uint64(0), cur.offset)
+	assert.Equal(t, uint64(0), head.offset)
 }
 
 func TestStream_TryWrite_One(t *testing.T) {
@@ -22,14 +22,14 @@ func TestStream_TryWrite_One(t *testing.T) {
 
 	num, _, _ := buf.TryWrite([]byte{1})
 
-	assert.Equal(t, uint32(1), num)
+	assert.Equal(t, uint64(1), num)
 	assert.Equal(t, []byte{1}, buf.Data())
 
 	tail, cur, head, _ := buf.Snapshot()
 
-	assert.Equal(t, uint32(0), tail.offset)
-	assert.Equal(t, uint32(0), cur.offset)
-	assert.Equal(t, uint32(1), head.offset)
+	assert.Equal(t, uint64(0), tail.offset)
+	assert.Equal(t, uint64(0), cur.offset)
+	assert.Equal(t, uint64(1), head.offset)
 }
 
 
@@ -37,43 +37,43 @@ func TestStream_TryWrite_ToCapacity(t *testing.T) {
 	buf := NewStream(4)
 
 	num, _, _ := buf.TryWrite([]byte{1,2,3,4})
-	assert.Equal(t, uint32(4), num)
+	assert.Equal(t, uint64(4), num)
 	assert.Equal(t, []byte{1, 2, 3, 4}, buf.Data())
 
 	tail, cur, head, _ := buf.Snapshot()
-	assert.Equal(t, uint32(0), tail.offset)
-	assert.Equal(t, uint32(0), cur.offset)
-	assert.Equal(t, uint32(4), head.offset)
+	assert.Equal(t, uint64(0), tail.offset)
+	assert.Equal(t, uint64(0), cur.offset)
+	assert.Equal(t, uint64(4), head.offset)
 }
 
 func TestStream_TryWrite_BeyondCapacity(t *testing.T) {
 	buf := NewStream(4)
 
 	num, _, _ := buf.TryWrite([]byte{1,2,3,4})
-	assert.Equal(t, uint32(4), num)
+	assert.Equal(t, uint64(4), num)
 	assert.Equal(t, []byte{1, 2, 3, 4}, buf.Data())
 
 	num, _, _ = buf.TryWrite([]byte{1,2,3,4})
-	assert.Equal(t, uint32(0), num)
+	assert.Equal(t, uint64(0), num)
 	assert.Equal(t, []byte{1, 2, 3, 4}, buf.Data())
 
 	tail, cur, head, _ := buf.Snapshot()
-	assert.Equal(t, uint32(0), tail.offset)
-	assert.Equal(t, uint32(0), cur.offset)
-	assert.Equal(t, uint32(4), head.offset)
+	assert.Equal(t, uint64(0), tail.offset)
+	assert.Equal(t, uint64(0), cur.offset)
+	assert.Equal(t, uint64(4), head.offset)
 }
 
 func TestStream_TryRead_Empty(t *testing.T) {
 	buf := NewStream(4)
 
 	start, num, _ := buf.TryRead([]byte{0}, false)
-	assert.Equal(t, uint32(0), start.offset)
-	assert.Equal(t, uint32(0), num)
+	assert.Equal(t, uint64(0), start.offset)
+	assert.Equal(t, uint64(0), num)
 
 	tail, cur, head, _ := buf.Snapshot()
-	assert.Equal(t, uint32(0), tail.offset)
-	assert.Equal(t, uint32(0), cur.offset)
-	assert.Equal(t, uint32(0), head.offset)
+	assert.Equal(t, uint64(0), tail.offset)
+	assert.Equal(t, uint64(0), cur.offset)
+	assert.Equal(t, uint64(0), head.offset)
 }
 
 func TestStream_TryRead_One(t *testing.T) {
@@ -83,14 +83,14 @@ func TestStream_TryRead_One(t *testing.T) {
 
 	out := make([]byte, 1)
 	start, num, _ := buf.TryRead(out, false)
-	assert.Equal(t, uint32(0), start.offset)
-	assert.Equal(t, uint32(1), num)
+	assert.Equal(t, uint64(0), start.offset)
+	assert.Equal(t, uint64(1), num)
 	assert.Equal(t, []byte{1}, out)
 
 	tail, cur, head, _ := buf.Snapshot()
-	assert.Equal(t, uint32(0), tail.offset)
-	assert.Equal(t, uint32(1), cur.offset)
-	assert.Equal(t, uint32(4), head.offset)
+	assert.Equal(t, uint64(0), tail.offset)
+	assert.Equal(t, uint64(1), cur.offset)
+	assert.Equal(t, uint64(4), head.offset)
 }
 
 func TestStream_Wrap(t *testing.T) {
@@ -100,17 +100,17 @@ func TestStream_Wrap(t *testing.T) {
 
 	out := make([]byte, 2)
 	start, num, _ := buf.TryRead(out, true)
-	assert.Equal(t, uint32(0), start.offset)
-	assert.Equal(t, uint32(2), num)
+	assert.Equal(t, uint64(0), start.offset)
+	assert.Equal(t, uint64(2), num)
 	assert.Equal(t, []byte{1,2}, out)
 
 	// the next write should wrap
 	buf.TryWrite([]byte{5,6})
 
 	tail, cur, head, _ := buf.Snapshot()
-	assert.Equal(t, uint32(2), tail.offset)
-	assert.Equal(t, uint32(2), cur.offset)
-	assert.Equal(t, uint32(6), head.offset)
+	assert.Equal(t, uint64(2), tail.offset)
+	assert.Equal(t, uint64(2), cur.offset)
+	assert.Equal(t, uint64(6), head.offset)
 
 	assert.Equal(t, []byte{3,4,5,6}, buf.Data())
 }
